Prepare for moving objc2-foundation
