+++ objc2-foundation/src/array.rs
@@ -9,10 +9,11 @@ use objc2::runtime::{Class, Object};
 use objc2::Message;
 use objc2::{msg_send, msg_send_id};
 
-use crate::{
+use super::{
     NSCopying, NSEnumerator, NSFastEnumeration, NSFastEnumerator, NSMutableArray, NSMutableCopying,
-    NSObject, NSRange, __inner_extern_class,
+    NSObject, NSRange,
 };
+use crate::__inner_extern_class;
 
 __inner_extern_class! {
     /// TODO
