+++ objc2-foundation/src/mutable_attributed_string.rs
@@ -3,7 +3,8 @@ use core::fmt;
 use objc2::rc::{DefaultId, Id, Owned, Shared};
 use objc2::{msg_send, msg_send_id};
 
-use crate::{extern_class, NSAttributedString, NSCopying, NSMutableCopying, NSObject, NSString};
+use super::{NSAttributedString, NSCopying, NSMutableCopying, NSObject, NSString};
+use crate::extern_class;
 
 extern_class! {
     /// A mutable string that has associated attributes.
