+++ objc2-foundation/src/mutable_string.rs
@@ -6,7 +6,8 @@ use core::str;
 use objc2::rc::{DefaultId, Id, Owned, Shared};
 use objc2::{msg_send, msg_send_id};
 
-use crate::{extern_class, NSCopying, NSMutableCopying, NSObject, NSString};
+use super::{NSCopying, NSMutableCopying, NSObject, NSString};
+use crate::extern_class;
 
 extern_class! {
     /// A dynamic plain-text Unicode string object.
