+++ objc2-foundation/src/object.rs
@@ -5,7 +5,8 @@ use objc2::rc::{DefaultId, Id, Owned, Shared};
 use objc2::runtime::{Class, Object};
 use objc2::{class, msg_send, msg_send_bool, msg_send_id};
 
-use crate::{NSString, __inner_extern_class};
+use super::NSString;
+use crate::__inner_extern_class;
 
 __inner_extern_class! {
     @__inner
