+++ objc2-foundation/src/value.rs
@@ -13,7 +13,8 @@ use objc2::rc::{DefaultId, Id, Shared};
 use objc2::Encode;
 use objc2::{msg_send, msg_send_id};
 
-use super::{NSCopying, NSObject, __inner_extern_class};
+use super::{NSCopying, NSObject};
+use crate::__inner_extern_class;
 
 __inner_extern_class! {
     // `T: Eq` bound to prevent `NSValue<f32>` from being `Eq`
